fix predicate fees
