+++ docs/src/cookbook/transfer-all-assets.md
@@ -10,13 +10,15 @@ Lets quickly go over the setup:
 
 We prepare two wallets with randomized addresses. Next, we want one of our wallets to have some random assets, so we set them up with `setup_multiple_assets_coins()`. Having created the coins, we can start a provider and assign it to the previously created wallets.
 
-Transactions require us to define input and output coins. Let's assume we do not know the assets owned by `wallet_1`. We retrieve its balances, i.e. tuples consisting of a string representing the asset id and the respective amount. This lets us use the helpers `get_asset_inputs_for_amount()`, `get_asset_outputs_for_amount()` to create the appropriate inputs and outputs:
+Transactions require us to define input and output coins. Let's assume we do not know the assets owned by `wallet_1`. We retrieve its balances, i.e. tuples consisting of a string representing the asset id and the respective amount. This lets us use the helpers `get_asset_inputs_for_amount()`, `get_asset_outputs_for_amount()` to create the appropriate inputs and outputs.
+
+For the sake of simplicity, we avoid transferring the base asset so we don't have to worry about transaction fees:
 
 ```rust,ignore
 {{#include ../../../examples/cookbook/src/lib.rs:transfer_multiple_inout}}
 ```
 
-All that is left is to build the transaction with the provider's helper `build_transfer_transaction()`, have `wallet_1` sign it, and we can send it. Checking that the wallets balances are empty lets us confirm that we have indeed transferred all assets:
+All that is left is to build the transaction with the helper `build_transfer_transaction()`, have `wallet_1` sign it, and we can send it. We confirm this by checking the number of balances present in the receiving wallet and their amount:
 
 ```rust,ignore
 {{#include ../../../examples/cookbook/src/lib.rs:transfer_multiple_transaction}}
